+++ api/constants.go
@@ -1,5 +1,15 @@
 package api
 
+const (
+	HeaderAppName    = "X-App-Name"
+	HeaderAppVersion = "X-App-Version"
+
+	HeaderWebhookToken = "X-Webhook-Token"
+
+	HeaderContentType                = "Content-Type"
+	HeaderContentTypeApplicationJson = "application/json"
+)
+
 // UpdateState state of an update
 type UpdateState string