+++ commons/constants.go
@@ -1,0 +1,5 @@
+package commons
+
+const (
+	Version = "3.0.2"
+)