+++ server/api_handler_action.go
@@ -225,6 +225,6 @@ func (h *actionHandler) delete(c *gin.Context) {
 		return
 	}
 
-	c.Header(headerContentType, headerContentTypeApplicationJson)
+	c.Header(api.HeaderContentType, api.HeaderContentTypeApplicationJson)
 	c.Status(http.StatusNoContent)
 }