+++ server/api_handler_action_invocation.go
@@ -97,6 +97,6 @@ func (h *actionInvocationHandler) delete(c *gin.Context) {
 		return
 	}
 
-	c.Header(headerContentType, headerContentTypeApplicationJson)
+	c.Header(api.HeaderContentType, api.HeaderContentTypeApplicationJson)
 	c.Status(http.StatusNoContent)
 }