+++ server/api_handler_auth.go
@@ -1,6 +1,7 @@
 package server
 
 import (
+	"git.myservermanager.com/varakh/upda/api"
 	"github.com/gin-gonic/gin"
 	"net/http"
 )
@@ -13,6 +14,6 @@ func newAuthHandler() *authHandler {
 }
 
 func (h *authHandler) login(c *gin.Context) {
-	c.Header(headerContentType, headerContentTypeApplicationJson)
+	c.Header(api.HeaderContentType, api.HeaderContentTypeApplicationJson)
 	c.Status(http.StatusNoContent)
 }