+++ server/constants_app.go
@@ -1,6 +1,5 @@
 package server
 
 const (
-	name    = "upda"
-	version = "3.0.2"
+	name = "upda"
 )